+++ pages/schedules.tsx
@@ -1,8 +1,8 @@
 import GetStartedBroadcast from 'components/GetStartedBroadcast'
 import HeadTitle from 'components/HeadTitle'
 import { AppLayout, MainContainer } from 'components/layouts'
-import Schedules from 'components/screens/Schedules'
-import Welcome from 'components/screens/welcomeS'
+import Schedules from 'components/lazy/schedules'
+import Intro from 'components/screens/intro'
 import { useSessionValue } from 'hooks/session'
 
 const title = 'Schedules'
@@ -11,7 +11,7 @@ export default function SchedulesPage() {
   const session = useSessionValue()
 
   if (session === '') {
-    return <Welcome />
+    return <Intro />
   }
 
   return (
